+++ Tinodios/MessageViewController.swift
@@ -26,12 +26,20 @@ protocol MessageDisplayLogic: AnyObject {
     func togglePreviewBar(with preview: NSAttributedString?, onAction action: PendingPreviewAction)
 }
 
+// Pending message is the one the user is either  replying to or forwarding.
+protocol PendingMessagePreviewDelegate: AnyObject {
+    // Calculates size for preview attributed string.
+    func pendingPreviewMessageSize(forMessage msg: NSAttributedString) -> CGSize
+    // Cancels preview.
+    func dismissPendingMessagePreview()
+}
+
 class MessageViewController: UIViewController {
     // Other controllers may send these notification to MessageViewController which will execute corresponding send message action.
     public static let kNotificationSendAttachment = "SendAttachment"
 
     // MARK: static parameters
-    private enum Constants {
+    enum Constants {
         /// Size of the avatar in the nav bar in small state.
         static let kNavBarAvatarSmallState: CGFloat = 32
 
@@ -133,14 +141,14 @@ class MessageViewController: UIViewController {
 
     /// The `forwardMessageBar` is used as the `inputAccessoryView` in the view controller
     /// (for forwarded messages only).
-    private lazy var forwardMessageBar: ForwardMessageBar = {
+    lazy var forwardMessageBar: ForwardMessageBar = {
         let view = ForwardMessageBar()
         view.autoresizingMask = .flexibleHeight
         return view
     }()
 
     /// Avatar in the NavBar
-    private lazy var navBarAvatarView: AvatarWithOnlineIndicator = {
+    lazy var navBarAvatarView: AvatarWithOnlineIndicator = {
         let avatarIcon = AvatarWithOnlineIndicator()
         let tapGestureRecognizer = UITapGestureRecognizer(target: self, action: #selector(navBarAvatarTapped(tapGestureRecognizer:)))
         avatarIcon.isUserInteractionEnabled = true
@@ -149,7 +157,7 @@ class MessageViewController: UIViewController {
     }()
 
     /// Call button in NavBar
-    private lazy var navBarCallBtn: UIBarButtonItem = {
+    lazy var navBarCallBtn: UIBarButtonItem = {
         return UIBarButtonItem(
             image: UIImage(systemName: "phone",
                            withConfiguration:UIImage.SymbolConfiguration(pointSize: 16, weight: .light)),
@@ -157,7 +165,7 @@ class MessageViewController: UIViewController {
     }()
 
     /// Pinned messages view.
-    private lazy var pinnedMessagesView: PinnedMessagesView = {
+    lazy var pinnedMessagesView: PinnedMessagesView = {
         let panel = PinnedMessagesView()
         panel.topicName = self.topicName
         self.view.addSubview(panel)
@@ -178,12 +186,12 @@ class MessageViewController: UIViewController {
     weak var collectionView: MessageView!
     private var collectionViewBottomAnchor: NSLayoutConstraint!
     /// Button [GO to latest message].
-    weak var goToLatestButton: UIButton!
+    private weak var goToLatestButton: UIButton!
     private var goToLatestButtonBottomAnchor: NSLayoutConstraint!
 
 
     var interactor: (MessageBusinessLogic & MessageDataStore)?
-    private let refreshControl = UIRefreshControl()
+    let refreshControl = UIRefreshControl()
 
     // MARK: properties
 
@@ -212,18 +220,15 @@ class MessageViewController: UIViewController {
     // * Database message id -> message offset.
     var messageDbIdIndex: [Int64: Int] = [:]
 
-    // Cache of message cell sizes. Calculation of cell sizes is heavy. Caching the result to improve scrolling performance.
-    // var cellSizeCache: [CGSize?] = []
-
     var isInitialLayout = true
 
     // Highlight this cell when scroll finishes (after the user tapped on a quote).
     var highlightCellAtPathAfterScroll: IndexPath?
 
     // Size of the present update batch.
-    private var updateBatchSize = 0
+    var updateBatchSize = 0
     // Last message received timestamp - for tracking batches.
-    private var lastMessageReceived = Date.distantPast
+    var lastMessageReceived = Date.distantPast
 
     // The two below indicate whether to send typing notifications and read receipts.
     // Determined by the user specified account notifications settings.
@@ -369,9 +374,9 @@ class MessageViewController: UIViewController {
         collectionView.translatesAutoresizingMaskIntoConstraints = false
         self.collectionViewBottomAnchor = collectionView.bottomAnchor.constraint(equalTo: view.bottomAnchor, constant: -(inputAccessoryView?.frame.height ?? 0))
         NSLayoutConstraint.activate([
-            collectionView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: Constants.kPinnedMessagesViewHeight),
+            //collectionView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: Constants.kPinnedMessagesViewHeight),
             // collectionView.topAnchor.constraint(equalTo: pinnedMessagesView.bottomAnchor),
-            // collectionView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
+            collectionView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
             self.collectionViewBottomAnchor,
             collectionView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor),
             collectionView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor)
@@ -629,243 +634,13 @@ class MessageViewController: UIViewController {
     }
 }
 
-// Methods for updating title area and refreshing messages.
-extension MessageViewController: MessageDisplayLogic {
-    private func showInvitationDialog() {
-        guard self.presentedViewController == nil else { return }
-        let attrs = [ NSAttributedString.Key.font: UIFont.systemFont(ofSize: 20.0) ]
-        let title = NSAttributedString(string: NSLocalizedString("New Chat", comment: "View title"), attributes: attrs)
-        let alert = UIAlertController(
-            title: nil,
-            message: NSLocalizedString("You are invited to start a new chat. What would you like?", comment: "Call to action"),
-            preferredStyle: .actionSheet)
-        alert.setValue(title, forKey: "attributedTitle")
-        alert.addAction(UIAlertAction(
-            title: NSLocalizedString("Accept", comment: "Invite reaction button"), style: .default,
-            handler: { _ in
-                self.interactor?.acceptInvitation()
-        }))
-        alert.addAction(UIAlertAction(
-            title: NSLocalizedString("Ignore", comment: "Invite reaction button"), style: .default,
-            handler: { _ in
-                self.interactor?.ignoreInvitation()
-        }))
-        alert.addAction(UIAlertAction(
-            title: NSLocalizedString("Block", comment: "Invite reaction button"), style: .default,
-            handler: { _ in
-                self.interactor?.blockTopic()
-        }))
-        self.present(alert, animated: true)
-    }
-
-    func switchTopic(topic: String?) {
-        topicName = topic
-    }
-
-    func updateTitleBar(pub: TheCard?, online: Bool?, deleted: Bool) {
-        assert(Thread.isMainThread)
-        self.navigationItem.title = pub?.fn ?? NSLocalizedString("Undefined", comment: "Undefined chat name")
-        navBarAvatarView.set(pub: pub, id: topicName, online: online, deleted: deleted)
-        navBarAvatarView.bounds = CGRect(x: 0, y: 0, width: Constants.kNavBarAvatarSmallState, height: Constants.kNavBarAvatarSmallState)
-
-        navBarAvatarView.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-                navBarAvatarView.heightAnchor.constraint(equalToConstant: Constants.kNavBarAvatarSmallState),
-                navBarAvatarView.widthAnchor.constraint(equalTo: navBarAvatarView.heightAnchor)
-            ])
-        var items = [UIBarButtonItem(customView: navBarAvatarView)]
-        let webrtc = Cache.tinode.getServerParam(for: "iceServers") != nil
-        if let t = self.topic, t.isP2PType, webrtc {
-            items.append(self.navBarCallBtn)
-        }
-        self.navigationItem.setRightBarButtonItems(items, animated: false)
-    }
-
-    func displayPinnedMessages(pins: [Int], selected: Int) {
-        assert(Thread.isMainThread)
-        if pins.isEmpty {
-            print("displayPinnedMessages -- empty \(pins)")
-            pinnedMessagesView.isHidden = true
-        } else {
-            print("Shown \(pins)")
-            pinnedMessagesView.topicName = self.topicName
-            pinnedMessagesView.pins = pins
-            pinnedMessagesView.isHidden = false
-        }
-    }
-
-    func setOnline(online: Bool?) {
-        assert(Thread.isMainThread)
-        navBarAvatarView.online = online
-    }
-
-    func runTypingAnimation() {
-        assert(Thread.isMainThread)
-        navBarAvatarView.presentTypingAnimation(steps: 30)
-    }
-
-    func reloadAllMessages() {
-        assert(Thread.isMainThread)
-        collectionView.reloadSections(IndexSet(integer: 0))
-    }
-
-    func displayChatMessages(messages: [StoredMessage], _ scrollToMostRecentMessage: Bool) {
-        assert(Thread.isMainThread)
-        guard collectionView != nil else { return }
-        let oldData = self.messages
-        let newData: [StoredMessage] = messages
-
-        // Both empty: no change.
-        guard !oldData.isEmpty || !newData.isEmpty else { return }
-
-        self.messageSeqIdIndex = newData.enumerated().reduce([Int: Int]()) { (dict, item) -> [Int: Int] in
-            var dict = dict
-            dict[item.element.seqId] = item.offset
-            return dict
-        }
-        self.messageDbIdIndex = newData.enumerated().reduce([Int64: Int]()) { (dict, item) -> [Int64: Int] in
-            var dict = dict
-            dict[item.element.msgId] = item.offset
-            return dict
-        }
-
-        // Update batch stats.
-        let now = Date()
-        let delta = newData.count - oldData.count
-        if now.millisecondsSince1970 > self.lastMessageReceived.millisecondsSince1970 + Constants.kUpdateBatchTimeDeltaThresholdMs {
-            self.updateBatchSize = delta
-        } else {
-            self.updateBatchSize += delta
-        }
-        self.lastMessageReceived = now
-
-        if oldData.isEmpty || newData.isEmpty || self.updateBatchSize > Constants.kUpdateBatchFullRefreshThreshold {
-            self.messages = newData
-            collectionView.reloadSections(IndexSet(integer: 0))
-            collectionView.layoutIfNeeded()
-            if scrollToMostRecentMessage {
-                collectionView.scrollToBottom()
-            }
-        } else {
-            // Get indexes of inserted and deleted items.
-            let diff = Utils.diffMessageArray(sortedOld: oldData, sortedNew: newData)
-
-            // Each insertion or deletion may change the appearance of the preceeding and following messages.
-            // Calculate indexes of all items which need to be updated.
-            var refresh: [Int] = []
-            for index in diff.mutated {
-                if index > 0 {
-                    // Refresh the preceeding item.
-                    refresh.append(index - 1)
-                }
-                if index < newData.count - 1 {
-                    // Refresh the following item.
-                    refresh.append(index + 1)
-                }
-                if index < newData.count {
-                    refresh.append(index)
-                }
-            }
-            // Ensure uniqueness of values. No need to reload newly inserted values.
-            // The app will crash if the same index is marked as removed and refreshed. Which seems
-            // to be an Apple bug because removed index is against the old array, refreshed against the new.
-            refresh = Array(Set(refresh).subtracting(Set(diff.inserted)))
-
-            if !diff.inserted.isEmpty || !diff.removed.isEmpty {
-                collectionView.performBatchUpdates({ () -> Void in
-                    self.messages = newData
-                    if diff.removed.count > 0 {
-                        collectionView.deleteItems(at: diff.removed.map { IndexPath(item: $0, section: 0) })
-                    }
-                    if diff.inserted.count > 0 {
-                        collectionView.insertItems(at: diff.inserted.map { IndexPath(item: $0, section: 0) })
-                    }
-                }, completion: nil)
-            }
-            if !refresh.isEmpty {
-                collectionView.performBatchUpdates({ () -> Void in
-                    self.messages = newData
-                    self.collectionView.reloadItems(at: refresh.map { IndexPath(item: $0, section: 0) })
-                    self.collectionView.layoutIfNeeded()
-                    if scrollToMostRecentMessage {
-                        self.collectionView.scrollToBottom()
-                    }
-                }, completion: nil)
-            }
-        }
-    }
-
-    func reloadMessages(fromSeqId loId: Int, toSeqId hiId: Int) {
-        assert(Thread.isMainThread)
-        guard self.collectionView != nil else { return }
-        let hiIdUpper = hiId + 1
-        let rowIds = (loId..<hiIdUpper).map { self.messageSeqIdIndex[$0] }.filter { $0 != nil }
-        self.collectionView.reloadItems(at: rowIds.map { IndexPath(item: $0!, section: 0) })
-    }
-
-    func updateProgress(forMsgId msgId: Int64, progress: Float) {
-        assert(Thread.isMainThread)
-        if let index = self.messageDbIdIndex[msgId],
-            let cell = self.collectionView.cellForItem(at: IndexPath(row: index, section: 0)) as? MessageCell {
-            cell.progressView.setProgress(progress)
-        }
-    }
-
-    func applyTopicPermissions(withError err: Error? = nil) {
-        assert(Thread.isMainThread)
-        // Make sure the view is visible.
-        guard self.isViewLoaded && ((self.view?.window) != nil) else { return }
-
-        if !(self.topic?.isReader ?? false) || err != nil {
-            self.collectionView.showNoAccessOverlay(withMessage: err?.localizedDescription)
-        } else {
-            self.collectionView.removeNoAccessOverlay()
-        }
-
-        let publishingForbidden = !(self.topic?.isWriter ?? false) || err != nil
-        // No "W" permission. Replace input field with a message "Not available".
-        self.sendMessageBar.toggleNotAvailableOverlay(visible: publishingForbidden)
-        if publishingForbidden {
-            // Dismiss all pending messages.
-            self.togglePreviewBar(with: nil)
-            self.interactor?.dismissPendingMessage()
-        }
-        // The peer is missing either "W" or "R" permissions. Show "Peer's messaging is disabled" message.
-        if let acs = self.topic?.peer?.acs, let missing = acs.missing {
-            self.sendMessageBar.togglePeerMessagingDisabled(visible: acs.isJoiner(for: .want) && (missing.isReader || missing.isWriter))
-        }
-        // We are offered to join a chat.
-        if let acs = self.topic?.accessMode, acs.isJoiner(for: Acs.Side.given) && (acs.excessive?.description.contains("RW") ?? false) {
-            self.showInvitationDialog()
-        }
-    }
-
-    func endRefresh() {
-        assert(Thread.isMainThread)
-        self.refreshControl.endRefreshing()
-    }
-
-    func dismissVC() {
-        assert(Thread.isMainThread)
-        self.navigationController?.popViewController(animated: true)
-        self.dismiss(animated: true)
-    }
-
-    func togglePreviewBar(with preview: NSAttributedString?, onAction action: PendingPreviewAction = .none) {
-        if preview == nil {
-            isForwardingMessage = false
-        }
-        if isForwardingMessage {
-            self.forwardMessageBar.togglePendingPreviewBar(with: preview)
-        } else {
-            self.sendMessageBar.togglePendingPreviewBar(withMessage: preview, onAction: action)
-        }
-        self.reloadInputViews()
-    }
-}
-
 // Methods for filling message with content and layout out message subviews.
 extension MessageViewController: UICollectionViewDataSource {
+
+    func numberOfSections(in: UICollectionView) -> Int {
+        print("numberOfSections")
+        return 1
+    }
 
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         self.collectionView.toggleNoMessagesNote(on: messages.isEmpty)
@@ -920,6 +695,18 @@ extension MessageViewController: UICollectionViewDataSource {
         bubbleDecorator(for: message, at: indexPath)(cell.containerView)
 
         return cell
+    }
+
+    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
+        print("Header requested \(kind) == \(UICollectionView.elementKindSectionHeader)")
+        if kind == UICollectionView.elementKindSectionHeader {
+            print("Dequeing section header")
+             let sectionHeader = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: String(describing: TempSectionHeader.self), for: indexPath) as! TempSectionHeader
+             sectionHeader.label.text = "This is a section header"
+             return sectionHeader
+        }
+
+        return UICollectionReusableView()
     }
 
     private func configureCell(cell: MessageCell, with message: Message, at indexPath: IndexPath) {
@@ -1305,14 +1092,6 @@ extension MessageViewController: MessageViewLayoutDelegate {
     }
 }
 
-// Pending message is the one the user is either  replying to or forwarding.
-protocol PendingMessagePreviewDelegate: AnyObject {
-    // Calculates size for preview attributed string.
-    func pendingPreviewMessageSize(forMessage msg: NSAttributedString) -> CGSize
-    // Cancels preview.
-    func dismissPendingMessagePreview()
-}
-
 extension MessageViewController: PendingMessagePreviewDelegate {
     func pendingPreviewMessageSize(forMessage msg: NSAttributedString) -> CGSize {
         return self.textSizeHelper.computeSize(for: msg, within: CGFloat.infinity)
@@ -1358,3 +1137,11 @@ extension MessageViewController: UICollectionViewDelegate {
         cell.highlightAnimated(withDuration: 4.0)
     }
 }
+
+extension MessageViewController: UICollectionViewDelegateFlowLayout {
+    public func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, referenceSizeForHeaderInSection section: Int) -> CGSize {
+        print("Header size requested")
+        return CGSize(width: collectionView.frame.width, height: 50)
+    }
+}
+
