+++ Tinodios/widgets/PinnedMessagesView.xib
@@ -31,7 +31,7 @@
                     </connections>
                 </button>
                 <view userInteractionEnabled="NO" contentMode="scaleAspectFit" translatesAutoresizingMaskIntoConstraints="NO" id="fKR-5v-8cl" userLabel="Page Indicator" customClass="DotSelectorImageView" customModule="Tinodios" customModuleProvider="target">
-                    <rect key="frame" x="41.666666666666664" y="0.0" width="23.999999999999993" height="65"/>
+                    <rect key="frame" x="41.666666666666664" y="2" width="23.999999999999993" height="61"/>
                     <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                     <accessibility key="accessibilityConfiguration">
                         <accessibilityTraits key="traits" notEnabled="YES"/>
@@ -41,20 +41,20 @@
                     </constraints>
                 </view>
                 <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="jVo-2x-tyr" customClass="PagerView" customModule="Tinodios" customModuleProvider="target">
-                    <rect key="frame" x="65.666666666666657" y="0.0" width="355.33333333333337" height="65"/>
-                    <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                    <rect key="frame" x="65.666666666666657" y="2" width="353.33333333333337" height="61"/>
+                    <color key="backgroundColor" systemColor="secondarySystemFillColor"/>
                 </view>
             </subviews>
-            <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+            <color key="backgroundColor" systemColor="secondarySystemFillColor"/>
             <constraints>
                 <constraint firstItem="jVo-2x-tyr" firstAttribute="leading" secondItem="fKR-5v-8cl" secondAttribute="trailing" id="1wc-QN-fub"/>
                 <constraint firstItem="Q3g-pt-nlo" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" constant="8" id="33a-ho-dfa"/>
-                <constraint firstItem="jVo-2x-tyr" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="6lM-ha-ErP"/>
-                <constraint firstItem="fKR-5v-8cl" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="JBR-Bw-DYt"/>
-                <constraint firstAttribute="trailing" secondItem="jVo-2x-tyr" secondAttribute="trailing" id="M0E-jb-UgW"/>
-                <constraint firstAttribute="bottom" secondItem="fKR-5v-8cl" secondAttribute="bottom" id="U3F-pG-N2C"/>
+                <constraint firstItem="jVo-2x-tyr" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" constant="2" id="6lM-ha-ErP"/>
+                <constraint firstItem="fKR-5v-8cl" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" constant="2" id="JBR-Bw-DYt"/>
+                <constraint firstAttribute="trailing" secondItem="jVo-2x-tyr" secondAttribute="trailing" constant="2" id="M0E-jb-UgW"/>
+                <constraint firstAttribute="bottom" secondItem="fKR-5v-8cl" secondAttribute="bottom" constant="2" id="U3F-pG-N2C"/>
                 <constraint firstItem="fKR-5v-8cl" firstAttribute="leading" secondItem="Q3g-pt-nlo" secondAttribute="trailing" constant="8" symbolic="YES" id="enC-dE-A3U"/>
-                <constraint firstAttribute="bottom" secondItem="jVo-2x-tyr" secondAttribute="bottom" id="yHY-gv-GdR"/>
+                <constraint firstAttribute="bottom" secondItem="jVo-2x-tyr" secondAttribute="bottom" constant="2" id="yHY-gv-GdR"/>
                 <constraint firstItem="Q3g-pt-nlo" firstAttribute="centerY" secondItem="iN0-l3-epB" secondAttribute="centerY" id="zRI-Md-MLb"/>
             </constraints>
             <freeformSimulatedSizeMetrics key="simulatedDestinationMetrics"/>
@@ -63,6 +63,9 @@
     </objects>
     <resources>
         <image name="xmark.circle" catalog="system" width="128" height="123"/>
+        <systemColor name="secondarySystemFillColor">
+            <color red="0.47058823529411764" green="0.47058823529411764" blue="0.50196078431372548" alpha="0.16" colorSpace="custom" customColorSpace="sRGB"/>
+        </systemColor>
         <systemColor name="systemBackgroundColor">
             <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
         </systemColor>
