+++ Tinodios/widgets/TempSectionHeader.swift
@@ -1,0 +1,34 @@
+//
+//  TempSectionHeader.swift
+//  Tinodios
+//
+//  Copyright © 2023 Tinode LLC. All rights reserved.
+//
+
+import UIKit
+
+public class TempSectionHeader: UICollectionReusableView {
+     var label: UILabel = {
+         let label: UILabel = UILabel()
+         label.textColor = .black
+         label.font = UIFont.systemFont(ofSize: 16, weight: .semibold)
+         label.sizeToFit()
+         return label
+     }()
+
+     override init(frame: CGRect) {
+         super.init(frame: frame)
+
+         addSubview(label)
+         backgroundColor = .red
+
+         label.translatesAutoresizingMaskIntoConstraints = false
+         label.topAnchor.constraint(equalTo: self.topAnchor).isActive = true
+         label.leftAnchor.constraint(equalTo: self.leftAnchor, constant: 20).isActive = true
+         label.rightAnchor.constraint(equalTo: self.rightAnchor).isActive = true
+    }
+
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
